+++ visitor-app/src/visitors/pages/QrError.js
