+++ visitor-app/src/visitors/pages/Thanks.js
